feat: definitional viz ranges, fractional fps, visible AOI outline
- Viz defaults: NDVI/EVI/NDWI/NDMI -> -1..1 (their definitional range); LST family
  (lst/lst_smw/lst_sharp/lst_modis) -> min -10C so sub-zero winter surface temps
  aren't clipped. Per-run viz.min/max still overrides.
- fps is now float end-to-end (config/api/gui), so 0.5 fps (2 s/frame) works; animate()
  defaults to a slower fps=2, and the Colab notebook exposes FPS (default 1).
- draw_region draws a dark casing under the amber core, so the AOI outline stays
  visible on hot (yellow/red) LST frames where the amber alone vanished.

+++ gee_animation/config.py
@@ -32,7 +32,7 @@ class RunConfig:
     viz_min: float
     viz_max: float
     palette: list[str]
-    fps: int
+    fps: float          # frames/sec; fractional allowed (e.g. 0.5 = 2 s per frame)
     scale: float
     dimensions: int
     # Render CRS (from render.crs). None => EPSG:4326 (plate carrée). "auto" => UTM
@@ -101,7 +101,7 @@ class RunConfig:
                 # explicit [] is preserved so validate() still rejects it); the
                 # composite default is None -> [].
                 palette=list(_p) if (_p := viz.get("palette", d_pal)) is not None else [],
-                fps=int(render["fps"]),
+                fps=float(render["fps"]),
                 scale=float(render["scale"]),
                 dimensions=int(render["dimensions"]),
                 crs=render.get("crs"),
